requirements-gemini.txt
google-cloud-texttospeech
pydub>=0.25.1
